feat: LINE & AREA fire; by Kalliser
New feature that allows to compute subtargets for a LINE or AREA between to targets.
Adds a new selection in the navbar to choose between POINT, LINE or AREA fire modes.
Displays subtargets that can be selected by clicking them.
Updated the footer, changing UI when LINE or AREA fire is selected.